refactor context parser & init()
